Update fitur update stok

+++ database/migrations/2025_11_08_140947_create_stoks_table.php
@@ -6,27 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('stoks', function (Blueprint $table) {
             $table->id();
             $table->string('nama_bahan');
             $table->string('satuan');
-            $table->integer('sisa_awal')->default(0);
-            $table->integer('sisa_sekarang')->default(0);
-            $table->integer('jumlah_keluar')->default(0);
+            $table->decimal('sisa_awal', 10, 2)->default(0);
+            $table->decimal('sisa_sekarang', 10, 2)->default(0);
+            $table->decimal('jumlah_keluar', 10, 2)->default(0);
             $table->date('tanggal_masuk')->nullable();
             $table->text('keterangan')->nullable();
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('stoks');
